+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.0.0 (HEAD)
 
+* Allow including attachment module to non-ActiveRecord::Base with activerecord plugin (janko-m)
+
 * Remove deprecated "restore_cached" alias for restore_cached_data plugin (janko-m)
 
 * Remove deprecated "delete_uploaded" alias for delete_raw plugin (janko-m)